+++ src/angle.rs
@@ -1,22 +1,23 @@
 //! Generic angles.
 
 use approx::{AbsDiffEq, RelativeEq, UlpsEq};
-use bytemuck::{Pod, Zeroable};
+use bytemuck::{Pod, TransparentWrapper, Zeroable};
 use core::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Rem, RemAssign, Sub, SubAssign};
-use num_traits::{Float, NumAssignOps};
 
 use crate::{
-    bindings::Quat, peel, prelude::*, scalar::FloatScalar, traits::marker::PodValue, Scalar, Unit,
-    Vector3,
+    bindings::Quat, peel, scalar::FloatScalar, traits::marker::PodValue, Scalar, Unit, Vector3,
 };
 
 /// Angle in radians.
 #[repr(transparent)]
-#[derive(Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, bytemuck::TransparentWrapper)]
-pub struct Angle<T = f32> {
+#[derive(Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord)]
+pub struct Angle<T: Scalar = f32> {
     /// Angle in radians.
     pub radians: T,
 }
+unsafe impl<T: Scalar> Zeroable for Angle<T> {}
+unsafe impl<T: Scalar> Pod for Angle<T> {}
+unsafe impl<T: Scalar> TransparentWrapper<T> for Angle<T> {}
 
 /// Strongly typed angle constants.
 ///
@@ -87,7 +88,7 @@ macro_rules! forward_to_float_as_angle {
 }
 
 /// Convenience trait to create strongly typed angles from floats.
-pub trait FloatAngleExt: num_traits::Float + AngleConsts + PodValue {
+pub trait FloatAngleExt: FloatScalar + AngleConsts + PodValue {
     forward_to_float_as_angle!(
         #[doc = "asin()"]
         fn asin(self) -> Angle<Self>
@@ -134,9 +135,6 @@ impl<T: Scalar + AngleConsts> core::fmt::Debug for Angle<T> {
     }
 }
 
-unsafe impl<T: Zeroable> Zeroable for Angle<T> {}
-unsafe impl<T: Pod> Pod for Angle<T> {}
-
 impl<T> Unit for Angle<T>
 where
     T: Scalar + AngleConsts,
@@ -148,10 +146,7 @@ where
     }
 }
 
-impl<T> AbsDiffEq for Angle<T>
-where
-    T: AbsDiffEq,
-{
+impl<T: Scalar> AbsDiffEq for Angle<T> {
     type Epsilon = T::Epsilon;
 
     #[inline]
@@ -170,10 +165,7 @@ where
     }
 }
 
-impl<T> UlpsEq for Angle<T>
-where
-    T: UlpsEq,
-{
+impl<T: FloatScalar> UlpsEq for Angle<T> {
     #[inline]
     fn default_max_ulps() -> u32 {
         T::default_max_ulps()
@@ -190,10 +182,7 @@ where
     }
 }
 
-impl<T> RelativeEq for Angle<T>
-where
-    T: RelativeEq,
-{
+impl<T: FloatScalar> RelativeEq for Angle<T> {
     #[inline]
     fn default_max_relative() -> Self::Epsilon {
         T::default_max_relative()
@@ -248,7 +237,7 @@ macro_rules! forward_to_unit_as_self {
     };
 }
 
-impl<T: Float> Angle<T> {
+impl<T: FloatScalar> Angle<T> {
     /// Angle from radians.
     #[inline]
     #[must_use]
@@ -303,7 +292,7 @@ impl<T: Float> Angle<T> {
         fn max(self, other: Self) -> Self);
 }
 
-impl<T: Float + AngleConsts> Angle<T> {
+impl<T: FloatScalar + AngleConsts> Angle<T> {
     /// 2π
     pub const CIRCLE: Self = Self::TAU;
     /// π
@@ -312,7 +301,7 @@ impl<T: Float + AngleConsts> Angle<T> {
     pub const QUARTER_CIRCLE: Self = Self::FRAG_PI_2;
 }
 
-impl<T> From<T> for Angle<T> {
+impl<T: Scalar> From<T> for Angle<T> {
     fn from(radians: T) -> Self {
         Angle { radians }
     }
@@ -320,7 +309,7 @@ impl<T> From<T> for Angle<T> {
 
 macro_rules! forward_op_scalar {
     ($trait_name:ident, $op:ident) => {
-        impl<T: Float> $trait_name<T> for Angle<T> {
+        impl<T: FloatScalar> $trait_name<T> for Angle<T> {
             type Output = Self;
 
             fn $op(self, rhs: T) -> Self {
@@ -334,7 +323,7 @@ macro_rules! forward_op_scalar {
 
 macro_rules! forward_op_scalar_assign {
     ($trait_name:ident, $op:ident) => {
-        impl<T: Float + NumAssignOps> $trait_name<T> for Angle<T> {
+        impl<T: FloatScalar> $trait_name<T> for Angle<T> {
             fn $op(&mut self, rhs: T) {
                 self.radians.$op(rhs);
             }
@@ -344,7 +333,7 @@ macro_rules! forward_op_scalar_assign {
 
 macro_rules! forward_op_self {
     ($trait_name:ident, $op:ident) => {
-        impl<T: Float> $trait_name for Angle<T> {
+        impl<T: FloatScalar> $trait_name for Angle<T> {
             type Output = Self;
 
             fn $op(self, rhs: Self) -> Self {
@@ -358,7 +347,7 @@ macro_rules! forward_op_self {
 
 macro_rules! forward_op_self_assign {
     ($trait_name:ident, $op:ident) => {
-        impl<T: Float + NumAssignOps> $trait_name for Angle<T> {
+        impl<T: FloatScalar> $trait_name for Angle<T> {
             fn $op(&mut self, rhs: Self) {
                 self.radians.$op(rhs.radians);
             }
@@ -380,7 +369,7 @@ forward_op_self_assign!(AddAssign, add_assign);
 forward_op_self_assign!(SubAssign, sub_assign);
 forward_op_self_assign!(RemAssign, rem_assign);
 
-impl<T: Div<Output = T>> Div<Angle<T>> for Angle<T> {
+impl<T: Scalar> Div<Angle<T>> for Angle<T> {
     type Output = T;
 
     fn div(self, rhs: Angle<T>) -> Self::Output {
@@ -388,7 +377,7 @@ impl<T: Div<Output = T>> Div<Angle<T>> for Angle<T> {
     }
 }
 
-impl<T: AngleConsts> AngleConsts for Angle<T> {
+impl<T: FloatScalar + AngleConsts> AngleConsts for Angle<T> {
     const PI: Self = Angle { radians: T::PI };
     const TAU: Self = Angle { radians: T::TAU };
 
@@ -415,10 +404,7 @@ impl<T: AngleConsts> AngleConsts for Angle<T> {
     };
 }
 
-impl<T> Angle<T>
-where
-    T: FloatScalar,
-{
+impl<T: FloatScalar> Angle<T> {
     /// Create quaternion from this angle and an axis vector.
     #[inline]
     #[must_use]