+++ src/scalar.rs
@@ -4,7 +4,11 @@ use crate::traits::marker::PodValue;
 /// All types that can serve as components of a SIMD type in [`glam`].
 ///
 /// This is implemented for `f32`, `f64`, `i32`, `u32`, `i16`, `u16`, `i64`, and `u64`.
-pub trait Scalar:
+///
+/// The associated types `Vec2`, `Vec3`, and `Vec4` define how this crate's vector/point/size types are mapped to `glam`
+/// vectors. Vectorlike types using a [`T: Unit`](crate::Unit) go through `T::Scalar` to find which `glam` vector type
+/// to use. Since those types are must be bitwise layout-compatible, this trait is unsafe to implement.
+pub unsafe trait Scalar:
     // This bound is important because it implies `Pod`, which means that types
     // that are `#[repr(C)]` and only have `Scalar` members can safely implement
     // `Pod` as well.
@@ -13,24 +17,22 @@ pub trait Scalar:
     + core::fmt::Display
     + crate::Unit<Scalar = Self>
     + num_traits::NumOps
+    + num_traits::NumAssignOps
     + num_traits::NumCast
+    + num_traits::ConstOne
+    + num_traits::ConstZero
     + approx::AbsDiffEq<Epsilon = Self>
     + From<bool>
 {
-    /// The underlying 2D vector type for this scalar ([`glam::Vec2`],
+    /// The underlying 2D vector type for this scalar (like [`glam::Vec2`],
     /// [`glam::DVec2`], [`glam::IVec2`], or [`glam::UVec2`]).
     type Vec2: bindings::Vector2<Scalar = Self>;
-    /// The underlying 3D vector type for this scalar ([`glam::Vec3`],
+    /// The underlying 3D vector type for this scalar (like [`glam::Vec3`],
     /// [`glam::DVec3`], [`glam::IVec3`], or [`glam::UVec3`]).
     type Vec3: bindings::Vector3<Scalar = Self>;
-    /// The underlying 4D vector type for this scalar ([`glam::Vec4`],
+    /// The underlying 4D vector type for this scalar (like [`glam::Vec4`],
     /// [`glam::DVec4`], [`glam::IVec4`], or [`glam::UVec4`]).
     type Vec4: bindings::Vector4<Scalar = Self>;
-
-    /// Zero.
-    const ZERO: Self;
-    /// One.
-    const ONE: Self;
 
     /// Try casting to another scalar type.
     ///
@@ -48,14 +50,22 @@ pub trait Scalar:
 
 /// Signed scalar types (`i32`, `f32`, `f64`, etc.).
 pub trait SignedScalar:
-    Scalar<Vec2: bindings::SignedVector2, Vec3: bindings::SignedVector, Vec4: bindings::SignedVector>
+    Scalar<
+        Vec2: bindings::SignedVector2,
+        Vec3: bindings::SignedVector,
+        Vec4: bindings::SignedVector,
+    > + num_traits::Signed
 {
     /// Negative one.
     const NEG_ONE: Self;
 }
 
 /// Floating-point scalar types (`f32` and `f64`).
-pub trait FloatScalar:
+///
+/// The associated types `Mat2`, `Mat3`, and `Mat4` define how this crate's matrix types are mapped to `glam` matrices.
+/// Matrixlike types using a [`T: Unit`](crate::Unit) go through `T::Scalar` to find which `glam` matrix type to use.
+/// Since those types are must be bitwise layout-compatible, this trait is unsafe to implement.
+pub unsafe trait FloatScalar:
     Scalar<
         Vec2: bindings::FloatVector2<Scalar = Self>,
         Vec3: bindings::FloatVector3<Scalar = Self>,
@@ -71,35 +81,23 @@ pub trait FloatScalar:
     /// Negative infinity.
     const NEG_INFINITY: Self;
 
-    // Note: Hiding these because they are implementation details to aid type inference.
-    #[doc(hidden)]
+    /// The underlying 2x2 matrix type for this scalar ([`glam::Mat2`] or [`glam::DMat2`]).
     type Mat2: bindings::Matrix2<Self>;
-    #[doc(hidden)]
+    /// The underlying 3x3 matrix type for this scalar ([`glam::Mat3`] or [`glam::DMat3`]).
     type Mat3: bindings::Matrix3<Self>;
-    #[doc(hidden)]
+    /// The underlying 4x4 matrix type for this scalar ([`glam::Mat4`] or [`glam::DMat4`]).
     type Mat4: bindings::Matrix4<Self>;
-    #[doc(hidden)]
+    /// The underlying quaternion type for this scalar ([`glam::Quat`] or [`glam::DQuat`]).
     type Quat: bindings::Quat<Self>;
-
-    /// True if the number is not NaN and not infinity.
-    #[must_use]
-    fn is_finite(self) -> bool;
 }
 
 /// Integer scalar types.
-pub trait IntScalar: Scalar<Vec2 = Self::Vec2i, Vec3 = Self::Vec3i, Vec4 = Self::Vec4i> {
-    #[doc(hidden)]
-    type Vec2i: bindings::Vector2<Scalar = Self> + bindings::IntegerVector;
-    #[doc(hidden)]
-    type Vec3i: bindings::Vector3<Scalar = Self> + bindings::IntegerVector;
-    #[doc(hidden)]
-    type Vec4i: bindings::Vector4<Scalar = Self> + bindings::IntegerVector;
+pub trait IntScalar:
+    Scalar<Vec2: bindings::IntegerVector, Vec3: bindings::IntegerVector, Vec4: bindings::IntegerVector>
+{
 }
 
-impl Scalar for f32 {
-    const ZERO: Self = 0.0;
-    const ONE: Self = 1.0;
-
+unsafe impl Scalar for f32 {
     type Vec2 = glam::Vec2;
     type Vec3 = glam::Vec3;
     type Vec4 = glam::Vec4;
@@ -109,7 +107,7 @@ impl SignedScalar for f32 {
     const NEG_ONE: Self = -1.0;
 }
 
-impl FloatScalar for f32 {
+unsafe impl FloatScalar for f32 {
     type Mat2 = glam::Mat2;
     type Mat3 = glam::Mat3;
     type Mat4 = glam::Mat4;
@@ -117,16 +115,9 @@ impl FloatScalar for f32 {
     const NAN: Self = f32::NAN;
     const INFINITY: Self = f32::INFINITY;
     const NEG_INFINITY: Self = f32::NEG_INFINITY;
-
-    fn is_finite(self) -> bool {
-        num_traits::Float::is_finite(self)
-    }
 }
 
-impl Scalar for f64 {
-    const ZERO: Self = 0.0;
-    const ONE: Self = 1.0;
-
+unsafe impl Scalar for f64 {
     type Vec2 = glam::DVec2;
     type Vec3 = glam::DVec3;
     type Vec4 = glam::DVec4;
@@ -136,7 +127,7 @@ impl SignedScalar for f64 {
     const NEG_ONE: Self = -1.0;
 }
 
-impl FloatScalar for f64 {
+unsafe impl FloatScalar for f64 {
     type Mat2 = glam::DMat2;
     type Mat3 = glam::DMat3;
     type Mat4 = glam::DMat4;
@@ -144,16 +135,9 @@ impl FloatScalar for f64 {
     const NAN: Self = f64::NAN;
     const INFINITY: Self = f64::INFINITY;
     const NEG_INFINITY: Self = f64::NEG_INFINITY;
-
-    fn is_finite(self) -> bool {
-        num_traits::Float::is_finite(self)
-    }
 }
 
-impl Scalar for i16 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for i16 {
     type Vec2 = glam::I16Vec2;
     type Vec3 = glam::I16Vec3;
     type Vec4 = glam::I16Vec4;
@@ -163,10 +147,7 @@ impl SignedScalar for i16 {
     const NEG_ONE: Self = -1;
 }
 
-impl Scalar for i32 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for i32 {
     type Vec2 = glam::IVec2;
     type Vec3 = glam::IVec3;
     type Vec4 = glam::IVec4;
@@ -176,10 +157,7 @@ impl SignedScalar for i32 {
     const NEG_ONE: Self = -1;
 }
 
-impl Scalar for i64 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for i64 {
     type Vec2 = glam::I64Vec2;
     type Vec3 = glam::I64Vec3;
     type Vec4 = glam::I64Vec4;
@@ -189,28 +167,19 @@ impl SignedScalar for i64 {
     const NEG_ONE: Self = -1;
 }
 
-impl Scalar for u16 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for u16 {
     type Vec2 = glam::U16Vec2;
     type Vec3 = glam::U16Vec3;
     type Vec4 = glam::U16Vec4;
 }
 
-impl Scalar for u32 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for u32 {
     type Vec2 = glam::UVec2;
     type Vec3 = glam::UVec3;
     type Vec4 = glam::UVec4;
 }
 
-impl Scalar for u64 {
-    const ZERO: Self = 0;
-    const ONE: Self = 1;
-
+unsafe impl Scalar for u64 {
     type Vec2 = glam::U64Vec2;
     type Vec3 = glam::U64Vec3;
     type Vec4 = glam::U64Vec4;
@@ -251,15 +220,5 @@ mod tests {
             Some(core::f32::NEG_INFINITY)
         );
         assert_eq!(1.0f64.try_cast::<f32>(), Some(1.0));
-    }
-
-    #[test]
-    fn finite() {
-        assert!(!FloatScalar::is_finite(f32::NAN));
-        assert!(!FloatScalar::is_finite(f32::INFINITY));
-        assert!(!FloatScalar::is_finite(f32::NEG_INFINITY));
-        assert!(!FloatScalar::is_finite(f64::NAN));
-        assert!(!FloatScalar::is_finite(f64::INFINITY));
-        assert!(!FloatScalar::is_finite(f64::NEG_INFINITY));
     }
 }